refactor(trusty-analyze): split 6 files under 500-SLOC cap (#1195)
Behavior-preserving split of the 6 remaining over-cap production files in
trusty-analyze into thin facade + focused sibling submodules, mirroring the
merged ruby/core-review pattern. All public paths preserved via re-exports.

Language adapters (foo.rs -> foo/{mod,nodes,calls,tests}.rs):
- lang/adapters/typescript.rs (560) -> mod 305 / nodes 47 / calls 56
- lang/adapters/python.rs (602)     -> mod / nodes / calls / tests
- lang/adapters/go.rs (610)         -> mod / nodes / calls / tests
- lang/adapters/php.rs (642)        -> mod / nodes / calls / tests

Bespoke splits:
- main.rs (567 -> 401): serve/deep/review-pr handler bodies + OutputFormat
  extracted to commands/run.rs
- mcp/mod.rs (892 -> 415): pure helpers -> mcp/helpers.rs, HTTP verbs ->
  mcp/http_client.rs, dispatch tests -> mcp/tests.rs

.line-cap-allowlist.tsv: dropped the 6 now-resolved trusty-analyze entries.

Verified: cargo build/test (429+25+2 pass), clippy -D warnings clean,
fmt --check clean, check_line_cap.sh 0 violations.
