+++ crates/trusty-analyze/src/mcp/http_client.rs
@@ -1,0 +1,130 @@
+//! HTTP transport verbs for the MCP dispatcher.
+//!
+//! Why: every MCP tool ultimately performs one HTTP call against the analyzer
+//! daemon; the GET/POST/DELETE plumbing is identical error-handling boilerplate
+//! lifted out of the dispatcher so `mcp/mod.rs` stays under the 500-SLOC
+//! production cap (see #1195).
+//! What: `pub(super)` methods on [`AnalyzerMcpServer`] that issue a request,
+//! map transport failures to [`DispatchError::Transport`], and decode the JSON
+//! body (treating non-2xx as a transport error carrying the response body).
+//! Test: exercised by every per-tool handler test in `tests.rs` (which asserts
+//! the constructed URL surfaces in the transport error when the daemon is down).
+
+use super::{AnalyzerMcpServer, DispatchError};
+use serde_json::Value;
+
+impl AnalyzerMcpServer {
+    pub(super) async fn post_text(&self, path: &str, body: &str) -> Result<Value, DispatchError> {
+        let url = format!("{}{}", self.base_url, path);
+        let resp = self
+            .http
+            .post(&url)
+            .header("content-type", "text/x-patch")
+            .body(body.to_string())
+            .send()
+            .await
+            .map_err(|e| DispatchError::Transport(format!("POST {url}: {e}")))?;
+        let status = resp.status();
+        let body: Value = resp
+            .json()
+            .await
+            .map_err(|e| DispatchError::Transport(format!("decode {url}: {e}")))?;
+        if !status.is_success() {
+            return Err(DispatchError::Transport(format!(
+                "POST {url} returned {status}: {body}"
+            )));
+        }
+        Ok(body)
+    }
+
+    pub(super) async fn post_bytes(
+        &self,
+        path: &str,
+        body: Vec<u8>,
+    ) -> Result<Value, DispatchError> {
+        let url = format!("{}{}", self.base_url, path);
+        let resp = self
+            .http
+            .post(&url)
+            .header("content-type", "application/octet-stream")
+            .body(body)
+            .send()
+            .await
+            .map_err(|e| DispatchError::Transport(format!("POST {url}: {e}")))?;
+        let status = resp.status();
+        let body: Value = resp
+            .json()
+            .await
+            .map_err(|e| DispatchError::Transport(format!("decode {url}: {e}")))?;
+        if !status.is_success() {
+            return Err(DispatchError::Transport(format!(
+                "POST {url} returned {status}: {body}"
+            )));
+        }
+        Ok(body)
+    }
+
+    pub(super) async fn get(&self, path: &str) -> Result<Value, DispatchError> {
+        let url = format!("{}{}", self.base_url, path);
+        let resp = self
+            .http
+            .get(&url)
+            .send()
+            .await
+            .map_err(|e| DispatchError::Transport(format!("GET {url}: {e}")))?;
+        let status = resp.status();
+        let body: Value = resp
+            .json()
+            .await
+            .map_err(|e| DispatchError::Transport(format!("decode {url}: {e}")))?;
+        if !status.is_success() {
+            return Err(DispatchError::Transport(format!(
+                "GET {url} returned {status}: {body}"
+            )));
+        }
+        Ok(body)
+    }
+
+    pub(super) async fn post(&self, path: &str, body: &Value) -> Result<Value, DispatchError> {
+        let url = format!("{}{}", self.base_url, path);
+        let resp = self
+            .http
+            .post(&url)
+            .json(body)
+            .send()
+            .await
+            .map_err(|e| DispatchError::Transport(format!("POST {url}: {e}")))?;
+        let status = resp.status();
+        let body: Value = resp
+            .json()
+            .await
+            .map_err(|e| DispatchError::Transport(format!("decode {url}: {e}")))?;
+        if !status.is_success() {
+            return Err(DispatchError::Transport(format!(
+                "POST {url} returned {status}: {body}"
+            )));
+        }
+        Ok(body)
+    }
+
+    pub(super) async fn delete(&self, path: &str) -> Result<Value, DispatchError> {
+        let url = format!("{}{}", self.base_url, path);
+        let resp = self
+            .http
+            .delete(&url)
+            .send()
+            .await
+            .map_err(|e| DispatchError::Transport(format!("DELETE {url}: {e}")))?;
+        let status = resp.status();
+        let body: Value = resp
+            .json()
+            .await
+            .map_err(|e| DispatchError::Transport(format!("decode {url}: {e}")))?;
+        if !status.is_success() {
+            return Err(DispatchError::Transport(format!(
+                "DELETE {url} returned {status}: {body}"
+            )));
+        }
+        Ok(body)
+    }
+}
